midst of adding and updating pytorch notebooks
