+++ credoai/artifacts/__init__.py
@@ -7,6 +7,7 @@ them.
 """
 from .data.base_data import Data
 from .data.comparison_data import ComparisonData
+from .data.pytorch_data import PytorchData
 from .data.tabular_data import TabularData
 from .model.base_model import Model
 from .model.classification_model import ClassificationModel, DummyClassifier
