--- credoai/datasets/_constants.py
@@ -1,1 +1,0 @@
-_DOWNLOAD_DIRECTORY_NAME = ".credoai"
