+++ credoai/demo_assets/__init__.py
@@ -1,0 +1,9 @@
+"""
+Data, models and other static resources needed for CredoAI lens
+"""
+
+from ._fetch_censusincome import fetch_censusincome
+from ._fetch_creditdefault import fetch_creditdefault
+from ._fetch_testdata import fetch_testdata
+from ._fetch_credit_model import fetch_credit_model
+from ._fetch_pytorch_mnist_model import fetch_mnist_model
