+++ credoai/demo_assets/_constants.py
@@ -1,0 +1,1 @@
+_DOWNLOAD_DIRECTORY_NAME = ".credoai"
